+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Relocate the username/password in the create/edit and detail view screens. Refs UIU-1035.
 * Allow search by last name. Refs UIU-1706.
 * Add user's fees/fines permissions. Refs UIU-1918.
+* Allow cancel fee/fine as error only if any 'actions' has not been applied. Refs UIU-1894.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)